requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.3.0
channels==4.0.0
channels-redis==4.1.0
asgiref==3.7.2
redis==4.5.5
psycopg2-binary==2.9.7
dj-database-url==2.2.0
whitenoise==6.6.0
gunicorn==21.2.0
daphne==4.0.0
